chore: document localnet setup
