+++ package.json
@@ -11,7 +11,6 @@
     "deprecations:report": "tsx ./scripts/reportDeprecations.mts",
     "migrate:deploy:localnet": "pnpm -r --workspace-concurrency=1 run migrate:deploy:localnet",
     "localnet": "tsx ./scripts/start-localnet.mts",
-    "localnet:docker": "tsx ./scripts/start-localnet-docker.mts",
     "syncpack:list": "pnpm syncpack list-mismatches",
     "syncpack:fix": "pnpm syncpack fix-mismatches",
     "typecheck": "pnpm -r typecheck",
